Move shuffle array function from CreateWalletPage to helpers
